Separate `core` and `compatibility` pipelines
